remove reading.rs & relocate useful benches from it
